More heavy code reorg.  Getting close to the state described in #80.
