+++ hendrix/__init__.py
@@ -1,1 +1,3 @@
-__version__ = '1.0.1'
+# This milesone lists the tasks remaining before we release 1.0:
+# https://github.com/hangarunderground/hendrix/milestones/1.0%20ness
+__version__ = '0.9.0'
